Split out the data package
