+++ app/src/main/java/me/banes/chris/tivi/data/daos/UserDao.kt
@@ -14,12 +14,22 @@
  * limitations under the License.
  */
 
-package me.banes.chris.tivi.data
+package me.banes.chris.tivi.data.daos
 
+import android.arch.persistence.room.Dao
+import android.arch.persistence.room.Insert
+import android.arch.persistence.room.OnConflictStrategy
+import android.arch.persistence.room.Query
 import io.reactivex.Flowable
+import me.banes.chris.tivi.data.entities.TraktUser
 
-interface EntryDao<PC, in EC : Entry> {
-    fun entries(): Flowable<List<PC>>
-    fun insert(entry: EC): Long
-    fun deleteAll()
+@Dao
+interface UserDao {
+
+    @Query("SELECT * FROM users")
+    fun getTraktUser(): Flowable<TraktUser>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun insertUser(user: TraktUser): Long
+
 }
