+++ app/src/main/java/me/banes/chris/tivi/home/HomeFragmentViewModel.kt
@@ -19,7 +19,7 @@ package me.banes.chris.tivi.home
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
 import me.banes.chris.tivi.AppNavigator
-import me.banes.chris.tivi.data.TraktUser
+import me.banes.chris.tivi.data.entities.TraktUser
 import me.banes.chris.tivi.extensions.plusAssign
 import me.banes.chris.tivi.trakt.TraktManager
 import me.banes.chris.tivi.util.ReactiveLiveData
