+++ app/src/main/java/me/banes/chris/tivi/home/HomeNavigator.kt
@@ -16,7 +16,7 @@
 
 package me.banes.chris.tivi.home
 
-import me.banes.chris.tivi.data.TiviShow
+import me.banes.chris.tivi.data.entities.TiviShow
 
 interface HomeNavigator {
