+++ app/src/main/java/me/banes/chris/tivi/trakt/TraktManager.kt
@@ -24,7 +24,7 @@ import io.reactivex.Maybe
 import io.reactivex.subjects.BehaviorSubject
 import me.banes.chris.tivi.AppNavigator
 import me.banes.chris.tivi.calls.UserMeCall
-import me.banes.chris.tivi.data.TraktUser
+import me.banes.chris.tivi.data.entities.TraktUser
 import me.banes.chris.tivi.extensions.edit
 import me.banes.chris.tivi.util.AppRxSchedulers
 import net.openid.appauth.AuthState
